refactor: fix parameter types

+++ system/Images/Handlers/BaseHandler.php
@@ -559,12 +559,12 @@ abstract class BaseHandler implements ImageHandlerInterface
         [$cropWidth, $cropHeight] = $this->calcAspectRatio($width, $height, $origWidth, $origHeight);
 
         if ($height === null) {
-            $height = ceil(($width / $cropWidth) * $cropHeight);
+            $height = (int) ceil(($width / $cropWidth) * $cropHeight);
         }
 
         [$x, $y] = $this->calcCropCoords($cropWidth, $cropHeight, $origWidth, $origHeight, $position);
 
-        return $this->crop($cropWidth, $cropHeight, $x, $y)->resize($width, $height);
+        return $this->crop($cropWidth, $cropHeight, (int) $x, (int) $y)->resize($width, $height);
     }
 
     /**
